feat: change version modules to match major releases names
