+++ src/tractusx_sdk/dataspace/controllers/connector/jupiter/contract_definition_controller.py
@@ -21,22 +21,21 @@
 #################################################################################
 
 from .dma_controller import DmaController
-from ..mixins import CrudDmaController
-from ....models.connector.v0_9_0 import PolicyModel
+from tractusx_sdk.dataspace.controllers.connector.utils.mixins import CrudDmaController
+from ....models.connector.jupiter import ContractDefinitionModel
 
 
-class PolicyController(CrudDmaController, DmaController):
+class ContractDefinitionController(CrudDmaController, DmaController):
     """
-    Concrete implementation of the PolicyController for the Connector v0.9.0 Data Management API.
+    Concrete implementation of the ContractDefinitionController for the Connector v0.9.0 Data Management API.
 
-    This class overrides the create and update methods in order to ensure the correct class types are used,
-    instead of the generic ones.
+    This class overrides the create and update methods in order to ensure the correct class types are used, instead of the generic ones.
     """
 
-    endpoint_url = "/v3/policydefinitions"
+    endpoint_url = "/v3/contractdefinitions"
 
-    def create(self, obj: PolicyModel, **kwargs):
+    def create(self, obj: ContractDefinitionModel, **kwargs):
         return super().create(obj, **kwargs)
 
-    def update(self, obj: PolicyModel, **kwargs):
+    def update(self, obj: ContractDefinitionModel, **kwargs):
         return super().update(obj, **kwargs)
