+++ src/tractusx_sdk/dataspace/controllers/connector/jupiter/contract_negotiation_controller.py
@@ -21,21 +21,24 @@
 #################################################################################
 
 from .dma_controller import DmaController
-from ..mixins import CrudDmaController
-from ....models.connector.v0_9_0 import ContractDefinitionModel
+from tractusx_sdk.dataspace.controllers.connector.utils.mixins import StatefulEntityDmaController
+from ....models.connector.jupiter import ContractNegotiationModel
 
 
-class ContractDefinitionController(CrudDmaController, DmaController):
+class ContractNegotiationController(StatefulEntityDmaController, DmaController):
     """
-    Concrete implementation of the ContractDefinitionController for the Connector v0.9.0 Data Management API.
+    Concrete implementation of the ContractNegotiationController for the Connector v0.9.0 Data Management API.
 
-    This class overrides the create and update methods in order to ensure the correct class types are used, instead of the generic ones.
+    This class overrides the create and terminate_by_id methods in order to ensure the correct class types are used, instead of the generic ones.
     """
 
-    endpoint_url = "/v3/contractdefinitions"
+    endpoint_url = "/v3/contractnegotiations"
 
-    def create(self, obj: ContractDefinitionModel, **kwargs):
+    def create(self, obj: ContractNegotiationModel, **kwargs):
         return super().create(obj, **kwargs)
 
-    def update(self, obj: ContractDefinitionModel, **kwargs):
-        return super().update(obj, **kwargs)
+    def terminate_by_id(self, oid: str, obj: ContractNegotiationModel, **kwargs):
+        return super().terminate_by_id(oid, obj, **kwargs)
+
+    def get_agreement_by_negotiation_id(self, oid: str, **kwargs):
+        return self.adapter.get(url=f"{self.endpoint_url}/{oid}/agreement", **kwargs)
