+++ src/tractusx_sdk/dataspace/controllers/connector/utils/decorators/controller_method.py
@@ -22,7 +22,7 @@
 
 import logging
 
-from ...controller import Controller
+from tractusx_sdk.dataspace.controllers.controller import Controller
 
 
 def controller_method(func):
