+++ src/tractusx_sdk/dataspace/models/connector/jupiter/asset_model.py
