+++ src/tractusx_sdk/dataspace/models/connector/jupiter/catalog_model.py
