+++ src/tractusx_sdk/dataspace/models/connector/jupiter/contract_negotiation_model.py
