+++ src/tractusx_sdk/dataspace/models/connector/jupiter/transfer_process_model.py
