+++ tests/dataspace/models/connector/test_model_factory_catalog.py
@@ -30,7 +30,7 @@ from tractusx_sdk.dataspace.models.connector.model_factory import ModelFactory
 
 class TestModelFactoryCatalog(unittest.TestCase):
     def setUp(self):
-        self.connector_version = "v0_9_0"
+        self.connector_version = "jupiter"
         self.counter_party_address = "https://counterparty.com"
         self.counter_party_id = "counterparty-id"
         self.context = {"key": "value"}
